Gitea Push
#my_printf.c

The function 'my_printf' itself is written in the file 'my_printf.c'. 
All necessary functions, which are used for 'my_printf' function are located in file 'functions.c'. 
There's also a file 'main.c'. It is created for your more convenient review. 

There is a 'Makefile', which assembles all files together. 
To make the compilation just type 'make' in your terminal. 
We run the code thereafter, with 'make run' command