Git Push: Bootcamp C Arc 01 quest04
#quest04

#ex01 - my_range.c
The function takes two integers, param_1 and param_2, and returns an array of integers.

Args: param_1: The first number in the range param_2: The maximum value of the range. Returns: An integer pointer to an array of integers.

my_range which returns a malloc'd array of integers
integer array should contain all values between min and max (Min included - max excluded.)
If min value is greater or equal to max's value, a null pointer is returned.
ex01 - my_strdup.c
The function my_strdup() takes a string as a parameter and returns a pointer to a newly allocated string that contains a copy of the argument string.

Args: param_1: The string to be duplicated. Returns: The function returns a pointer to the new string.

#ex02 - my_print_words_array.c
The function my_print_words_array takes a string_array as a parameter.

Args: param_1: a pointer to a string_array structure. Returns: Nothing

#ex03 - my_count_on_it.c
It iterates over the string array and counts the number of characters in each string.

Args: param_1: A string array. Returns: The integer array

#ex04 - my_join.c
The function my_join takes two parameters, a string array and a separator.

Args: param_1: a pointer to a string_array struct. param_2: The string to be inserted between each string in the array. Returns: The new string.

#ex05 - my_spaceship.c
The code takes in a string of commands and outputs the final position of the spaceship.

Args: param_1: A string of commands. Returns: A string with the x, y, and direction of the spaceship.